call flux_attr_get() for jobid
Since
* 485743d1 broker: clear job environment vars in rc scripts

commited to flux-core, there is not an environment variable FLUX_JOB_ID
set when a user is in an allocation but has not run a job, e.g.

flux mini run -n1 bash -c set | grep FLUX

However that allocation does still have a job ID, and it can be obtained
with flux_attr_get().